refactor: split mitm matching pattern tests
